refactor: move some functions
