+++ packages/mask/background/services/identity/index.ts
@@ -1,3 +1,8 @@
 export { createPersonaByPrivateKey } from './persona/create'
+export { mobile_queryPersonaRecordsFromIndexedDB } from './persona/query'
+export { deletePersona, logoutPersona, setupPersona } from './persona/update'
 export { signWithPersona, type SignRequest, type SignRequestResult, generateSignResult } from './persona/sign'
 export { exportPersonaMnemonicWords, exportPersonaPrivateKey } from './persona/backup'
+
+export { mobile_queryProfiles, mobile_queryProfileRecordFromIndexedDB } from './profile/query'
+export { updateProfileInfo, type UpdateProfileInfo, mobile_removeProfile } from './profile/update'
