+++ packages/mask/background/services/identity/persona/update.ts
@@ -1,0 +1,56 @@
+import type { PersonaIdentifier } from '@masknet/shared-base'
+import {
+    consistentPersonaDBWriteAccess,
+    queryPersonaDB,
+    detachProfileDB,
+    deletePersonaDB,
+    safeDeletePersonaDB,
+    updatePersonaDB,
+} from '../../../database/persona/db'
+
+export async function deletePersona(id: PersonaIdentifier, confirm: 'delete even with private' | 'safe delete') {
+    return consistentPersonaDBWriteAccess(async (t) => {
+        const d = await queryPersonaDB(id, t)
+        if (!d) return
+        for (const e of d.linkedProfiles) {
+            await detachProfileDB(e[0], t)
+        }
+        if (confirm === 'delete even with private') await deletePersonaDB(id, 'delete even with private', t)
+        else if (confirm === 'safe delete') await safeDeletePersonaDB(id, t)
+    })
+}
+
+export async function loginPersona(identifier: PersonaIdentifier) {
+    return consistentPersonaDBWriteAccess((t) =>
+        updatePersonaDB(
+            { identifier, hasLogout: false },
+            { linkedProfiles: 'merge', explicitUndefinedField: 'ignore' },
+            t,
+        ),
+    )
+}
+
+export async function logoutPersona(identifier: PersonaIdentifier) {
+    return consistentPersonaDBWriteAccess((t) =>
+        updatePersonaDB(
+            { identifier, hasLogout: true },
+            { linkedProfiles: 'merge', explicitUndefinedField: 'ignore' },
+            t,
+        ),
+    )
+}
+
+export async function setupPersona(id: PersonaIdentifier) {
+    return consistentPersonaDBWriteAccess(async (t) => {
+        const d = await queryPersonaDB(id, t)
+        if (!d) throw new Error('cannot find persona')
+        if (d.linkedProfiles.size === 0) throw new Error('persona should link at least one profile')
+        if (d.uninitialized) {
+            await updatePersonaDB(
+                { identifier: id, uninitialized: false },
+                { linkedProfiles: 'merge', explicitUndefinedField: 'ignore' },
+                t,
+            )
+        }
+    })
+}
