+++ packages/mask/src/database/Persona/helpers.ts
@@ -6,9 +6,6 @@ import {
     queryProfileDB,
     queryPersonaDB,
     queryPersonasDB,
-    detachProfileDB,
-    deletePersonaDB,
-    safeDeletePersonaDB,
     consistentPersonaDBWriteAccess,
     updatePersonaDB,
     queryProfilesPagedDB,
@@ -103,38 +100,6 @@ export async function queryPersonasWithQuery(query?: Parameters<typeof queryPers
     return _.map(personaRecordToPersona)
 }
 
-export async function deletePersona(id: PersonaIdentifier, confirm: 'delete even with private' | 'safe delete') {
-    return consistentPersonaDBWriteAccess(async (t) => {
-        const d = await queryPersonaDB(id, t)
-        if (!d) return
-        for (const e of d.linkedProfiles) {
-            await detachProfileDB(e[0], t)
-        }
-        if (confirm === 'delete even with private') await deletePersonaDB(id, 'delete even with private', t)
-        else if (confirm === 'safe delete') await safeDeletePersonaDB(id, t)
-    })
-}
-
-export async function loginPersona(identifier: PersonaIdentifier) {
-    return consistentPersonaDBWriteAccess((t) =>
-        updatePersonaDB(
-            { identifier, hasLogout: false },
-            { linkedProfiles: 'merge', explicitUndefinedField: 'ignore' },
-            t,
-        ),
-    )
-}
-
-export async function logoutPersona(identifier: PersonaIdentifier) {
-    return consistentPersonaDBWriteAccess((t) =>
-        updatePersonaDB(
-            { identifier, hasLogout: true },
-            { linkedProfiles: 'merge', explicitUndefinedField: 'ignore' },
-            t,
-        ),
-    )
-}
-
 export async function renamePersona(identifier: PersonaIdentifier, nickname: string) {
     const personas = await queryPersonasWithQuery({ nameContains: nickname })
     if (personas.length > 0) {
@@ -144,21 +109,6 @@ export async function renamePersona(identifier: PersonaIdentifier, nickname: str
     return consistentPersonaDBWriteAccess((t) =>
         updatePersonaDB({ identifier, nickname }, { linkedProfiles: 'merge', explicitUndefinedField: 'ignore' }, t),
     )
-}
-
-export async function setupPersona(id: PersonaIdentifier) {
-    return consistentPersonaDBWriteAccess(async (t) => {
-        const d = await queryPersonaDB(id, t)
-        if (!d) throw new Error('cannot find persona')
-        if (d.linkedProfiles.size === 0) throw new Error('persona should link at least one profile')
-        if (d.uninitialized) {
-            await updatePersonaDB(
-                { identifier: id, uninitialized: false },
-                { linkedProfiles: 'merge', explicitUndefinedField: 'ignore' },
-                t,
-            )
-        }
-    })
 }
 
 export async function queryPersonaByProfile(i: ProfileIdentifier) {
